Fixing build system + cross platformability
Former-commit-id: d382166500afba63de430433605288b1eb1cc64c
Former-commit-id: aaa1b69e5bcb5452d27fc9dcb1b49c01338ae481 [formerly 5d7e1436c1525b4c4d27e53e0b46a57ce61c9d3d]
Former-commit-id: 398bc7721db46bc432037298806d009b4a892c4e

+++ modules/renderer/source/hip/hiprenderer/shader/shadervar.d
@@ -199,7 +199,7 @@ struct ShaderVar
         return ShaderVar.create(t, varName, &dRef, UniformType.floating_array, dRef.sizeof, dRef[0].sizeof, true);
     }
 
-    protected static ShaderVar* create(ShaderTypes t, string varName, void* varData, UniformType type, ulong varSize, ulong singleSize, bool isDynamicArrayReference=false)
+    protected static ShaderVar* create(ShaderTypes t, string varName, void* varData, UniformType type, size_t varSize, size_t singleSize, bool isDynamicArrayReference=false)
     {
         import core.stdc.string : memcpy;
         import core.stdc.stdlib : malloc;
@@ -256,6 +256,7 @@ class ShaderVariablesLayout
     private string[] namesOrder;
     string name;
     ShaderTypes shaderType;
+    //Single block representation of variables content
     protected void* data;
     protected void* additionalData;
     protected bool isAdditionalAllocated;
@@ -367,7 +368,7 @@ class ShaderVariablesLayout
 
         return this;
     }
-    final ulong getLayoutSize(){return lastPosition;}
+    final size_t getLayoutSize(){return lastPosition;}
     final void setAdditionalData(void* d, bool isAllocated)
     {
         this.additionalData = d;
